Code in SelfStartingMonitoringView is not reusable in a non-spring application
